Integrated code for moving the usb missile launcher
The code for ensuring that the camera can be moved in the right
direction depending on where the face is located so that it can
track is integrated into main.py

tracker.py was mostly generated by ChatGPT and was used to get
the algorithm right. It will not be maintained any more.
It is added to the check-in to show how sections can be independently
developed and tested using ChatGPT before integrating them into
the main code.

Also imported the launcher into main.py. Moving the missile launcher/camera
itself is not done yet. Only pseudo movement is done. This enables
testing on faster laptop rather than raspberry pi which would be very slow.

For the proper integration we would need to do hardware modifications such
as mounting the raspberry pi camera module onto the USB missile launcher