Synchronize tracker.py with main.py for SOLID Principles Compliance
In this commit, tracker.py has been synchronized with main.py to lay
the groundwork for adherence to SOLID principles. This effort is also
geared towards making tracker.py independently testable, enhancing
modularity, and ensuring single responsibility within the codebase.

Key Modifications:
- Aligned the core functionalities and structures of tracker.py with
  those in main.py.
- Streamlined and refined code for better readability and maintenance.
- Ensured both files have a consistent approach towards handling image
  processing and face detection tasks.
- Prepared the codebase for future refactoring that aligns with SOLID
  principles, specifically focusing on the single responsibility and
  open/closed principles.

This update is a proactive step towards improving the code quality and
maintainability of the project, setting a strong foundation for further
enhancements and testing.

Addresses Issues: #38, #29

Testing Done:

Ran `python tracker.py` and made sure that the face recognition worked,
the virtual commands that printed the commands to move and launch the
T-Shirts worked.